python/requirements.txt
numpy==1.16.2
matplotlib==3.0.3
sounddevice==0.3.14
psutil==5.6.1
pyrtlsdr==0.2.91
scipy==1.5.2
numba==0.43.1